+++ lua/LUA/ak/io/ServerController.lua
@@ -6,15 +6,15 @@
 -- Use this class in XxxWebConnector to register JsonCollectors and commands
 local ServerController = require("ak.io.ServerController")
 --]] -- @author Andreas Kreuz
--- @release 0.10.0
+-- @release 0.10.1
 if AkDebugLoad then print("Loading ak.io.ServerController ...") end
 local AkWebServerIo = require("ak.io.AkWebServerIo")
 local AkCommandExecutor = require("ak.io.AkCommandExecutor")
 local os = require("os")
 
 local ServerController = {}
-ServerController.debug = AkDebugLoad or false
-ServerController.programVersion = "0.10.0"
+ServerController.debug = AkStartWithDebug or false
+ServerController.programVersion = "0.10.1"
 local json
 
 -- ACHTUNG: DIE VERWENDUNG ERFOLGT AUF EIGENE GEFAHR. ES IST GUT MÖGLICH,